+++ client/src/main/java/de/schlunzis/client/ClientLauncher.java
@@ -1,5 +1,6 @@
 package de.schlunzis.client;
 
+import de.schlunzis.client.fx.ClientApp;
 import javafx.application.Application;
 import net.rgielen.fxweaver.core.FxWeaver;
 import net.rgielen.fxweaver.spring.SpringFxWeaver;
@@ -9,7 +10,7 @@ import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
 public class ClientLauncher {
-    
+
     public static void main(String[] args) {
         Application.launch(ClientApp.class, args);
     }
