+++ client/src/main/java/de/schlunzis/client/fx/controller/MainMenuController.java
@@ -1,4 +1,4 @@
-package de.schlunzis.client.controller;
+package de.schlunzis.client.fx.controller;
 
 import de.schlunzis.common.messages.authentication.LogoutRequest;
 import javafx.event.ActionEvent;
