+++ client/src/main/java/de/schlunzis/client/fx/scene/SceneChangeEvent.java
@@ -1,0 +1,5 @@
+package de.schlunzis.client.fx.scene;
+
+public record SceneChangeEvent(Scene scene) {
+
+}
