--- client/src/main/java/de/schlunzis/client/scene/events/SceneChangeEvent.java
@@ -1,7 +1,0 @@
-package de.schlunzis.client.scene.events;
-
-import de.schlunzis.client.scene.Scene;
-
-public record SceneChangeEvent(Scene scene) {
-
-}
