+++ client/src/main/java/de/schlunzis/client/service/ISessionService.java
@@ -1,0 +1,10 @@
+package de.schlunzis.client.service;
+
+import de.schlunzis.common.User;
+
+import java.util.Optional;
+
+public interface ISessionService {
+
+    Optional<User> getCurrentUser();
+}
