+++ client/src/main/resources/de/schlunzis/client/fx/controller/ChatController.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import java.net.URL?>
 <VBox xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/17.0.2-ea"
-      fx:controller="de.schlunzis.client.controller.ChatController">
+      fx:controller="de.schlunzis.client.fx.controller.ChatController">
     <stylesheets>
         <URL value="@css/style.css"/>
         <URL value="@css/chat.css"/>
