+++ client/src/main/resources/de/schlunzis/client/fx/controller/css/style.css
